backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
python-multipart
httpx
pandas
scikit-learn
xgboost
joblib
